Changed modulesPath default value to repo root
better to do it this way and not wish it were not within the modules
directory later.

Also better to change it now while it is less than 12 hours old.

If I do not change this to the repository root, there is no possibility
of changing my mind later on where the helper modules will live if too
much feedback comes in that they should be moved