feat(makeWrapper): consolidated makeWrapper and makeWrapperNix
changed several options. The module system is a better interface than a
list of args, and it was hard for other implementations to parse the
list of args. So makeWrapperBase has been removed.

The escaping function options are better, with even more fine grained
control on the way soon.

the nix wrapper implementation is very useful, as if escapingFunction is
set to wlib.escapeShellArgWithEnv it will allow expansion of env vars at
runtime.